Add files via upload
Used Java's built in system to convert bytes to a string and a string back to bytes to get matching data in/out for server/client. Used type long to store the number of bits set/received to avoid negative rates. Added checks for hostname to be sure total length does not exceed 253 and that it contains valid characters.